+++ src/App.tsx
@@ -3,8 +3,8 @@ import { Outlet } from 'react-router-dom';
 import { RecoilRoot } from 'recoil';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ThemeProvider } from '@material-tailwind/react';
-import Header from '@components/Header';
-import Footer from '@components/Footer';
+import Header from '@components/Header/Header';
+import Footer from '@components/Footer/Footer';
 
 const queryClient = new QueryClient({
   defaultOptions: {
