+++ src/components/Modal/QuitModal.tsx
