+++ src/index.tsx
@@ -17,8 +17,8 @@ import GroupCreatePage from '@pages/GroupCreatePage';
 import ReportPage from '@pages/ReportPage';
 import PostEditPage from '@pages/PostEditPage';
 
-import MainLayout from '@components/MainLayout';
-import PageLayout from '@components/PageLayout';
+import MainLayout from '@components/Layout/MainLayout';
+import PageLayout from '@components/Layout/PageLayout';
 import App from './App';
 
 const router = createBrowserRouter([
